So we get much better output now. Still could probably do with some tidying
We now have an issue that only one of the last hops is shown.
This is because of the the mechanism to stop the search hits when any of the leaves reaches the goal. We need to circle that one last time to amend all the others without going any further.